Update daily aggregates to run all scalars in a single query
